Moved around a few things and implemented scrolling behavior for firefox.
