+++ gulp/src/common/browserSpecific.js
@@ -1,3 +1,1 @@
 
-// http://stackoverflow.com/questions/6927637/addeventlistener-in-internet-explorer
-export const isIE8 = !Boolean(window.addEventListener);
